pyproject.toml
[tool.poetry]
name = "searchform"
version = "0.1.0"
description = ""
authors = ["ByAvatarOff <tsp7439@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
djangorestframework = "^3.15.1"
requests = "^2.31.0"
tinydb = "^4.8.0"
python-dotenv = "^1.0.1"
pytest = "^8.2.0"
pytest-django = "^4.8.0"
pymongo = "^4.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
